mindsdb-main/requirements.txt
# Core MindsDB packages
mindsdb>=25.6.2.0
mindsdb[extra]>=25.6.2.0

# Core Dependencies
requests>=2.28.1
python-dotenv>=0.21.0

# Cloud Provider SDKs
boto3>=1.26.0
google-cloud-storage>=2.7.0
azure-storage-blob>=12.14.0
azure-identity>=1.12.0
oci>=2.90.0

# Development
pylint>=2.15.0
black>=22.10.0
isort>=5.10.1
mypy>=0.991

# Testing
pytest>=7.2.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.20.3
pytest-xdist>=3.1.0
pytest-timeout>=2.1.0
responses>=0.22.0

# Type checking
types-requests>=2.28.11
types-python-dateutil>=2.8.19
typing-extensions>=4.4.0

# Documentation
sphinx>=6.1.3
sphinx-rtd-theme>=1.2.0
sphinx-autodoc-typehints>=1.22.0
